Bulk changes.  Close to final working version

+++ redditrepostsleuth/core/model/image_index_api_result.py
@@ -18,5 +18,5 @@ class IndexSearchResult(BaseModel):
 
 class APISearchResults(BaseModel):
     total_searched: int = 0
-    total_search_time: int = 0
+    total_search_time: float = 0
     results: List[IndexSearchResult] = []
